refactor(test): split record_decision monolith and extract shared test doubles (#850)
- Splits 562-line test_record_decision.py into:
  - test_record_decision_validation.py (TestRecordDecisionValidation)
  - test_record_decision_handler.py (insert + resolution + regression guards)
  - test_record_decision_lib.py (UUID + resolve refs unit tests)
  - test_record_decision_helpers.py (shared mock factories + UUID constants)
- Extracts shared mock factories into test_record_decision_helpers.py
  with consistent naming (make_client, resolver_client)
- Deletes original monolith file

Closes #850
